Well, hooray, I found the problem. But, alas, It implies a refactor.
I was storing the resulting solution plan as a linear array but the solution needs to be a combination of an array at the base level and then a tree up each of the base level array members.. or, it needs to figure out based on the depth where it places itself in the triedMoves array. That may actually be easier.. if depth > 1 { /* place this tried move above the last one or something */ }

Well, saving before I continue.