Commit comes directly before trip to airport. Save game.
Built up basic search for solution. Fails when it needs to track open conditions and conflicts.

Did some review and got a better idea about what I need to implement. Took notes in doc/. a system to link predicates to predicates and moves to moves. Ordering is implied from pred to pred but not from move to move, so, we need to enforce ordering. All Edges are 1 to 1 except for reverse-flow predicate links which are one to many.
Assigning a link between two entities needs to create two links for two-way linking. Move links will go by assignment, predicate links by Array.prototype.push.
I have saved a bunch of material on this to refer to in docs/.